+++ meteor/client/lib/Components/LabelAndOverrides.tsx
@@ -4,10 +4,7 @@ import { objectPathGet } from '@sofie-automation/corelib/dist/lib'
 import React, { useCallback } from 'react'
 import { useTranslation } from 'react-i18next'
 import { ReadonlyDeep } from 'type-fest'
-import {
-	OverrideOpHelperForItemContents,
-	WrappedOverridableItemNormal,
-} from '../../../lib/collections/OverrideOpHelper'
+import { OverrideOpHelperForItemContents, WrappedOverridableItemNormal } from '../../ui/util/OverrideOpHelper'
 import { DropdownInputOption, findOptionByValue } from './DropdownInput'
 import { hasOpWithPath } from './util'
