+++ meteor/client/lib/forms/SchemaFormTable/ObjectTableOpHelper.tsx
@@ -1,5 +1,5 @@
 import { clone, joinObjectPathFragments, objectPathSet } from '@sofie-automation/corelib/dist/lib'
-import { OverrideOpHelperForItemContents, WrappedOverridableItem } from '../../../../lib/collections/OverrideOpHelper'
+import { OverrideOpHelperForItemContents, WrappedOverridableItem } from '../../../ui/util/OverrideOpHelper'
 
 /**
  * The OverrideOp system does not support tables of objects currently.
