+++ meteor/client/lib/forms/SchemaFormWithOverrides.tsx
@@ -1,10 +1,7 @@
 import { joinObjectPathFragments, literal } from '@sofie-automation/corelib/dist/lib'
 import React, { useMemo } from 'react'
 import { useTranslation } from 'react-i18next'
-import {
-	WrappedOverridableItemNormal,
-	OverrideOpHelperForItemContents,
-} from '../../../lib/collections/OverrideOpHelper'
+import { WrappedOverridableItemNormal, OverrideOpHelperForItemContents } from '../../ui/util/OverrideOpHelper'
 import { CheckboxControl } from '../Components/Checkbox'
 import { DropdownInputOption, DropdownInputControl } from '../Components/DropdownInput'
 import { FloatInputControl } from '../Components/FloatInput'
