+++ meteor/client/ui/RundownView/RundownRightHandControls.tsx
@@ -26,10 +26,7 @@ import { SegmentViewMode } from '../../lib/ui/icons/listView'
 import { RundownPlaylistId } from '@sofie-automation/corelib/dist/dataModel/Ids'
 import { MediaStatusPopUp } from './MediaStatusPopUp'
 import { MediaStatusIcon } from '../../lib/ui/icons/mediaStatus'
-import {
-	WrappedOverridableItemNormal,
-	getAllCurrentAndDeletedItemsFromOverrides,
-} from '../../../lib/collections/OverrideOpHelper'
+import { WrappedOverridableItemNormal, getAllCurrentAndDeletedItemsFromOverrides } from '../util/OverrideOpHelper'
 import { ObjectWithOverrides } from '@sofie-automation/corelib/dist/settings/objectWithOverrides'
 
 interface IProps {
