+++ meteor/client/ui/Settings/Studio/Devices/GenericSubDevices.tsx
@@ -6,7 +6,7 @@ import {
 	WrappedOverridableItem,
 	WrappedOverridableItemDeleted,
 	WrappedOverridableItemNormal,
-} from '../../../../../lib/collections/OverrideOpHelper'
+} from '../../../util/OverrideOpHelper'
 import { faCheck, faPencilAlt, faSync, faTrash } from '@fortawesome/free-solid-svg-icons'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { JSONBlob, JSONBlobParse, JSONSchema } from '@sofie-automation/blueprints-integration'
