+++ meteor/server/publications/packageManager/expectedPackages/util.ts
@@ -3,7 +3,7 @@ import { MappingExt, MappingsExt, StudioRouteSet } from '@sofie-automation/corel
 import { ReadonlyDeep } from 'type-fest'
 import { getActiveRoutes, getRoutedMappings } from '../../../../lib/collections/Studios'
 import { ObjectWithOverrides } from '@sofie-automation/corelib/dist/settings/objectWithOverrides'
-import { getAllCurrentAndDeletedItemsFromOverrides } from '../../../../lib/collections/OverrideOpHelper'
+import { getAllCurrentAndDeletedItemsFromOverrides } from '@sofie-automation/corelib/dist/overrideOpHelperBackend'
 
 type MappingExtWithOriginalName = MappingExt & { originalLayerName: string }
 type MappingsExtWithOriginalName = {
